Moved changes into session.rs to keep related functions together
