+++ llama-cpp-2/src/lib.rs
@@ -40,8 +40,7 @@ pub mod timing;
 pub mod token;
 pub mod token_type;
 
-pub use crate::context::session::LlamaStateSeqFlags;
-pub use crate::context::SeqState;
+pub use crate::context::session::{LlamaStateSeqFlags, SeqState};
 
 #[cfg(feature = "common")]
 pub(crate) fn status_is_ok(status: llama_cpp_sys_2::llama_rs_status) -> bool {
